refactor: getGetTiff so it can be used for other functions as well
