+++ fdm-calculator/tsconfig.json
@@ -18,7 +18,8 @@
         "strict": true,
         "noUnusedLocals": true,
         "noUnusedParameters": true,
-        "noFallthroughCasesInSwitch": true
+        "noFallthroughCasesInSwitch": true,
+        "baseUrl": "./src"
     },
     "include": ["src"]
 }
